refactor: remove utils package.
